+++ src/webservice.c
@@ -278,7 +278,7 @@ int callback_glewlwyd_user_auth (const struct _u_request * request, struct _u_re
               y_log_message(Y_LOG_LEVEL_ERROR, "callback_glewlwyd_user_auth - Error user_session_update (1)");
               response->status = 500;
             } else {
-              ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", config->cookie_secure, 0, config->cookie_same_site);
+              ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", (int)config->cookie_secure, 0, (int)config->cookie_same_site);
               y_log_message(Y_LOG_LEVEL_INFO, "Event - User '%s' authenticated with password", json_string_value(json_object_get(j_param, "username")));
             }
             o_free(session_uid);
@@ -304,7 +304,7 @@ int callback_glewlwyd_user_auth (const struct _u_request * request, struct _u_re
               y_log_message(Y_LOG_LEVEL_ERROR, "callback_glewlwyd_user_auth - Error user_session_update (2)");
               response->status = 500;
             } else {
-              ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", config->cookie_secure, 0, config->cookie_same_site);
+              ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", (int)config->cookie_secure, 0, (int)config->cookie_same_site);
             }
           } else if (check_result_value(j_result, G_ERROR_NOT_FOUND)) {
             response->status = 401;
@@ -335,7 +335,7 @@ int callback_glewlwyd_user_auth (const struct _u_request * request, struct _u_re
               y_log_message(Y_LOG_LEVEL_ERROR, "callback_glewlwyd_user_auth - Error user_session_update (3)");
               response->status = 500;
             } else {
-              ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", config->cookie_secure, 0, config->cookie_same_site);
+              ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", (int)config->cookie_secure, 0, (int)config->cookie_same_site);
               y_log_message(Y_LOG_LEVEL_INFO, "Event - User '%s' authenticated with scheme '%s/%s'", json_string_value(json_object_get(j_param, "username")), json_string_value(json_object_get(j_param, "scheme_type")), json_string_value(json_object_get(j_param, "scheme_name")));
             }
             o_free(session_uid);
@@ -368,7 +368,7 @@ int callback_glewlwyd_user_auth (const struct _u_request * request, struct _u_re
             y_log_message(Y_LOG_LEVEL_ERROR, "callback_glewlwyd_user_auth - Error user_session_update (4)");
             response->status = 500;
           } else {
-            ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", config->cookie_secure, 0, config->cookie_same_site);
+            ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", (int)config->cookie_secure, 0, (int)config->cookie_same_site);
             y_log_message(Y_LOG_LEVEL_INFO, "Event - User '%s' authenticated with scheme '%s/%s'", json_string_value(json_object_get(j_result, "username")), json_string_value(json_object_get(j_param, "scheme_type")), json_string_value(json_object_get(j_param, "scheme_name")));
           }
           o_free(session_uid);
@@ -667,9 +667,9 @@ int callback_glewlwyd_user_delete_session (const struct _u_request * request, st
         }
         if (json_array_size(json_object_get(j_session, "session")) == 1) {
           // Delete session cookie on the client browser
-          ulfius_add_same_site_cookie_to_response(response, config->session_key, "", expires, 0, config->cookie_domain, "/", config->cookie_secure, 0, config->cookie_same_site);
+          ulfius_add_same_site_cookie_to_response(response, config->session_key, "", expires, 0, config->cookie_domain, "/", (int)config->cookie_secure, 0, (int)config->cookie_same_site);
         } else {
-          ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", config->cookie_secure, 0, config->cookie_same_site);
+          ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", (int)config->cookie_secure, 0, (int)config->cookie_same_site);
         }
       } else {
         if (user_session_delete(config, session_uid, NULL) != G_OK) {
@@ -677,7 +677,7 @@ int callback_glewlwyd_user_delete_session (const struct _u_request * request, st
           response->status = 500;
         }
         // Delete session cookie on the client browser
-        ulfius_add_same_site_cookie_to_response(response, config->session_key, "", expires, 0, config->cookie_domain, "/", config->cookie_secure, 0, config->cookie_same_site);
+        ulfius_add_same_site_cookie_to_response(response, config->session_key, "", expires, 0, config->cookie_domain, "/", (int)config->cookie_secure, 0, (int)config->cookie_same_site);
       }
     }
     json_decref(j_session);
@@ -2371,7 +2371,7 @@ int callback_glewlwyd_user_get_profile (const struct _u_request * request, struc
       j_session = get_users_for_session(config, session_uid);
       if (check_result_value(j_session, G_OK)) {
         ulfius_set_json_body_response(response, 200, json_object_get(j_session, "session"));
-        ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", config->cookie_secure, 0, config->cookie_same_site);
+        ulfius_add_same_site_cookie_to_response(response, config->session_key, session_uid, expires, 0, config->cookie_domain, "/", (int)config->cookie_secure, 0, (int)config->cookie_same_site);
       } else if (check_result_value(j_session, G_ERROR_NOT_FOUND)) {
         response->status = 401;
       } else {
